requirements.txt
selenium>=3.141.0
webdriver-manager>=4.0.0
python-dotenv>=0.19.0
